requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mysql-connector==2.1.6
oauthlib==2.0.7
python-twitter==3.4.1
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
Werkzeug==0.14.1
